requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-genai
celery
redis
pydantic-settings
moviepy
